backend/requirements.txt
fastapi
uvicorn
openslide-python
scikit-image
pillow
numpy
python-multipart
scikit-learn 